reconfigure utils to modify and add flexible config for versions
