+++ requirements.txt
@@ -4,3 +4,4 @@ pandas>=1.4.2
 pyarrow>=10.0.1
 SALib>=1.4.5
 tqdm>=4.64.1
+pyyaml>=6.0.0
