+++ setup.py
@@ -30,5 +30,6 @@ setup(
         'pyarrow>=10.0.1',
         'SALib>1.4.5',
         'tqdm>=4.64.1',
+        'pyyaml>=6.0.0'
     ],
 )
