+++ statemodify/tests/test_utils.py
@@ -1,181 +1,57 @@
+import pkg_resources
 import unittest
 
-import pandas as pd
+import numpy as np
 
-import statemodify.utils as utils
+import statemodify.utils as utx
 
 
 class TestUtils(unittest.TestCase):
     """Tests to ensure functionality and stability of utils.py."""
 
-    COMP_RIGHT = "    test"
-    COMP_LEFT = "test    "
-    COMP_NONE = "test"
-    COMP_ZEROS = "0.10"
-    COMP_DATA = {'a': [0], 'b': ['alpha  '], 'c': [2.0]}
-    COMP_FILE = "/my/output/directory/file_scenario-test_sample-0.txt"
-    COMP_STR = "item   focus\n"
-    COMP_ADJ_DF_ADD = pd.DataFrame({"a": [10., 20.1], "b": [-1., -1.9]})
-    COMP_ADJ_DF_MULTIPLY = pd.DataFrame({"a": [10., 2.], "b": [-1., -0.2]})
-    FIELD_DICT = {"a": [], "b": [], "c": []}
-    COLUMN_WIDTHS = {"a": 1, "b": 7, "c": 3}
-    COLUMN_LIST = ["a", "b", "c"]
-    DATA_TYPES = {"a": int, "b": str, "c": float}
-    BAD_BOUNDS_LOW = [[-0.5, 1.0], [-0.7, 1.0]]
-    BAD_BOUNDS_HIGH = [[-0.5, 1.4], [-0.5, 1.0]]
-    BAD_BOUNDS_BOTH = [[-0.5, 1.4], [-0.7, 1.0]]
+    TEST_YAML = pkg_resources.resource_filename("statemodify", "data/ddm_data_specification.yml")
 
-    def test_set_alignment(self):
-        """Tests for the set_alignment function."""
+    def test_yaml_str(self):
+        """Test YAML return string."""
 
-        # ensure error is raised when value other than right or left is passed
-        with self.assertRaises(AssertionError):
-            s = utils.set_alignment(value="test",
-                                    n_spaces=0,
-                                    align="center")
+        d = utx.yaml_to_dict(TestUtils.TEST_YAML)
 
-        # check for left padding (right alignment) of four on value; total length of eight
-        s = utils.set_alignment(value="test",
-                                n_spaces=4,
-                                align="right")
+        self.assertEqual(str, type(d["comment_indicator"]), msg="Failure for YAML read to get string type.")
+        self.assertEqual("#", d["comment_indicator"], msg="Failure for YAML read to get string value.")
 
-        self.assertEqual(self.COMP_RIGHT, s)
-        self.assertEqual(len(self.COMP_RIGHT), len(s))
+    def test_yaml_dict_of_lists(self):
+        """Test YAML return dict of lists."""
 
-        # check for right padding (left alignment) of four on value; total length of eight
-        s = utils.set_alignment(value="test",
-                                n_spaces=4,
-                                align="left")
+        d = utx.yaml_to_dict(TestUtils.TEST_YAML)
 
-        self.assertEqual(self.COMP_LEFT, s)
-        self.assertEqual(len(self.COMP_LEFT), len(s))
+        self.assertEqual(dict, type(d["data_dict"]), msg="Failure for YAML read to get dict type.")
+        self.assertEqual(list, type(d["data_dict"]["prefix"]), msg="Failure for YAML read to get dict key type list.")
 
-        # check for no space of four on value; total length of eight
-        s = utils.set_alignment(value="test",
-                                n_spaces=0,
-                                align="left")
+    def test_yaml_dict_of_int(self):
+        """Test YAML return dict of int."""
 
-        self.assertEqual(self.COMP_NONE, s)
-        self.assertEqual(len(self.COMP_NONE), len(s))
+        d = utx.yaml_to_dict(TestUtils.TEST_YAML)
 
-        # check for n_spaces smaller than length of value of four; total length of eight
-        s = utils.set_alignment(value="test",
-                                n_spaces=-2,
-                                align="left")
+        self.assertEqual(dict, type(d["column_widths"]), msg="Failure for YAML read to get dict type.")
+        self.assertEqual(int, type(d["column_widths"]["prefix"]), msg="Failure for YAML read to get dict key type int.")
+        self.assertEqual(5, d["column_widths"]["prefix"], msg="Failure for YAML read to get dict key value int.")
 
-        self.assertEqual(self.COMP_NONE, s)
-        self.assertEqual(len(self.COMP_NONE), len(s))
+    def test_yaml_dict_of_types(self):
+        """Test YAML return dict of types."""
 
-    def test_pad_with_spaces(self):
-        """Tests for the pad_with_spaces function.  Majority of tests accounted for in child function."""
+        d = utx.yaml_to_dict(TestUtils.TEST_YAML)
 
-        # ensure error is raised when value length is greater than expected width
-        with self.assertRaises(TypeError):
-            s = utils.pad_with_spaces(value="  test  ",
-                                      expected_width=2,
-                                      align="right")
+        self.assertEqual(dict, type(d["data_types"]), msg="Failure for YAML read to get dict type.")
+        self.assertEqual(str, d["data_types"]["prefix"], msg="Failure for YAML read to get dict key type.")
+        self.assertEqual(np.float64, d["data_types"]["oct"], msg="Failure for YAML read to get dict key type.")
 
-    def test_add_zero_padding(self):
-        """Tests for the add_zero_padding function."""
+    def test_yaml_list(self):
+        """Test YAML return list."""
 
-        # ensure zeros are added to the desired precision
-        s = utils.add_zero_padding(x="0.1",
-                                   precision=2)
+        d = utx.yaml_to_dict(TestUtils.TEST_YAML)
 
-        self.assertEqual(self.COMP_ZEROS, s)
+        self.assertEqual(list, type(d["column_list"]), msg="Failure for YAML read to get list type.")
 
-    def test_populate_dict(self):
-        """Tests for the populate_dict function."""
 
-        # ensure parsing into dict is correct
-        d = utils.populate_dict(line="0alpha  2.0\n",
-                                field_dict=self.FIELD_DICT,
-                                column_widths=self.COLUMN_WIDTHS,
-                                column_list=self.COLUMN_LIST,
-                                data_types=self.DATA_TYPES)
-
-        self.assertEqual(self.COMP_DATA, d)
-
-    def test_construct_outfile_name(self):
-        """Tests for construct_outfile_name function."""
-
-        # ensure function produces what is expected
-        f = utils.construct_outfile_name(template_file="/my/template/file/file.txt",
-                                         output_directory="/my/output/directory",
-                                         scenario="test",
-                                         sample_id=0)
-
-        self.assertEqual(self.COMP_FILE, f)
-
-    def test_construct_data_string(self):
-        """Tests for construct_data_string function."""
-
-        data = utils.construct_data_string(df=pd.DataFrame({"a": ["item   "],
-                                                            "b": ["   focus  "]}),
-                                           column_list=["a", "b"],
-                                           column_widths={"a": 7, "b": 5},
-                                           column_alignment={"a": "left", "b": "right"})
-
-        self.assertEqual(self.COMP_STR, data)
-
-    def test_apply_adjustment_factor_add(self):
-        """Ensure add adjustment works."""
-
-        data = utils.apply_adjustment_factor(data_df=pd.DataFrame({"a": [10, 20],
-                                                                   "b": [-1, -2],
-                                                                   "c": [0, 1]}),
-                                             value_columns=["a", "b"],
-                                             query_field="c",
-                                             target_ids=[1],
-                                             factor=0.1,
-                                             factor_method="add")
-
-        pd.testing.assert_frame_equal(self.COMP_ADJ_DF_ADD, data)
-
-    def test_apply_adjustment_factor_multiply(self):
-        """Ensure multiply adjustment works."""
-
-        data = utils.apply_adjustment_factor(data_df=pd.DataFrame({"a": [10, 20],
-                                                                   "b": [-1, -2],
-                                                                   "c": [0, 1]}),
-                                             value_columns=["a", "b"],
-                                             query_field="c",
-                                             target_ids=[1],
-                                             factor=0.1,
-                                             factor_method="multiply")
-
-        pd.testing.assert_frame_equal(self.COMP_ADJ_DF_MULTIPLY, data)
-
-    def test_apply_adjustment_factor_error(self):
-        """Ensure error is raised for invalid factor_method."""
-
-        with self.assertRaises(KeyError):
-            data = utils.apply_adjustment_factor(data_df=pd.DataFrame({"a": [10, 20],
-                                                                       "b": [-1, -2],
-                                                                       "c": [0, 1]}),
-                                                 value_columns=["a", "b"],
-                                                 query_field="c",
-                                                 target_ids=[1],
-                                                 factor=0.1,
-                                                 factor_method="divide")
-
-    def test_validate_bounds_low(self):
-        """Raise error for low bound violation."""
-
-        with self.assertRaises(ValueError, msg="Low bounds validation failure."):
-            utils.validate_bounds(bounds_list=TestUtils.BAD_BOUNDS_LOW)
-
-    def test_validate_bounds_high(self):
-        """Raise error for high bound violation."""
-
-        with self.assertRaises(ValueError, msg="High bounds validation failure."):
-            utils.validate_bounds(bounds_list=TestUtils.BAD_BOUNDS_HIGH)
-
-    def test_validate_bounds_both(self):
-        """Raise error for both bound violation."""
-
-        with self.assertRaises(ValueError, msg="High and low bounds validation failure."):
-            utils.validate_bounds(bounds_list=TestUtils.BAD_BOUNDS_BOTH)
-
-    if __name__ == '__main__':
-        unittest.main()
+if __name__ == '__main__':
+    unittest.main()
